refactor(indexers): split VA index into its own file
